Sequential model transformations are broken. Replacing a pattern with replacement layers adds the replacement layers to the wrong part of the model. This basically ends up creating an incorrect model.
Adding a test to reproduce the error. The previous tests
were passing only because the model was solele composed
of matching layers. Further, `array.insert` in python
allows inserting into arbitrary index in an empty list.

This CL fixes the indexing bug. However, the replacement
code is quite complicated and should be simplified.

PiperOrigin-RevId: 312633923